+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 2.8.0
 
 * Skip boot2docker if Unix.
+* Added `permissionErrorTolerant` flag to deal with CircleCI btrfs errors.
 
 2.7.0
                                                                                                                                                Java